Transport configuration reworking

+++ zenoh/src/net/link/tcp/unicast.rs
@@ -129,7 +129,7 @@ impl LinkUnicastTrait for LinkUnicastTcp {
     }
 
     #[inline(always)]
-    fn get_mtu(&self) -> usize {
+    fn get_mtu(&self) -> u16 {
         *TCP_DEFAULT_MTU
     }
