+++ zenoh/src/net/link/udp/mod.rs
@@ -35,22 +35,22 @@ pub use unicast::*;
 //       Although in IPv6 it is possible to have UDP datagrams of size greater than 65,535 bytes via
 //       IPv6 Jumbograms, its usage in Zenoh is discouraged unless the consequences are very well
 //       understood.
-const UDP_MAX_MTU: usize = 65_507;
+const UDP_MAX_MTU: u16 = 65_507;
 
 #[cfg(any(target_os = "linux", target_os = "windows"))]
 // Linux default value of a maximum datagram size is set to UDP MAX MTU.
-const UDP_MTU_LIMIT: usize = UDP_MAX_MTU;
+const UDP_MTU_LIMIT: u16 = UDP_MAX_MTU;
 
 #[cfg(target_os = "macos")]
 // Mac OS X default value of a maximum datagram size is set to 9216 bytes.
-const UDP_MTU_LIMIT: usize = 9_216;
+const UDP_MTU_LIMIT: u16 = 9_216;
 
 #[cfg(not(any(target_os = "linux", target_os = "macos", target_os = "windows")))]
-const UDP_MTU_LIMIT: usize = 8_192;
+const UDP_MTU_LIMIT: u16 = 8_192;
 
 zconfigurable! {
     // Default MTU (UDP PDU) in bytes.
-    static ref UDP_DEFAULT_MTU: usize = UDP_MTU_LIMIT;
+    static ref UDP_DEFAULT_MTU: u16 = UDP_MTU_LIMIT;
     // Amount of time in microseconds to throttle the accept loop upon an error.
     // Default set to 100 ms.
     static ref UDP_ACCEPT_THROTTLE_TIME: u64 = 100_000;
